requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
amqp==5.3.1
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
by==0.0.4
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Delorean==1.0.0
Django==2.2.17
entsoe-py==0.6.16
et_xmlfile==2.0.0
h11==0.14.0
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.48
py==1.11.0
pycparser==2.22
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.3.4
pytest-assume==2.4.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-mock==1.12.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
